\name{concProbIns-package}
\alias{concProbIns-package}
\docType{package}
\title{
  Concordance probability for large datasets in an insurance pricing setting
}
\description{
  This package gives two estimation methods to calculate the concordance probability in an efficient way, both for frequency and severity data.
}
\details{
  \tabular{ll}{
    Package: \tab concProbIns\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
    License: \tab GPL (>= 2)\cr
  }
}

\references{Ponnet, J., Van Oirbeek, R., & Verdonck, T. (2021). Concordance Probability for Insurance Pricing Models. Risks, 9(10), 178.}

\author{
  Ponnet J., Van Oirbeek R., Verdonck T.\cr
  Maintainer: Ponnet J. <jolien.ponnet@kuleuven.be>
}

